Don't copy Keyboarding.dll.config file on Windows
Rename App.config file to Keyboarding.dll.config and remove it from .csproj file. Add buildinclude.xml file that
copies the Keyboarding.dll.config file when building on Linux. The config file isn't needed on Windows and Alistair
thought it would be better if it wouldn't polute the installer on Windows.