Allow to specify packages for NAnt build (FWNX-374)
Linux has the concept of packages (*.pc files) that specify the
location of libraries and other things needed for building. The
NAnt csc task accepts a <pkg-references> that lists packages that
should be referenced when building.

MonoDevelop lists the packages it finds but (in contrast to the
description in FWNX-374) no longer writes <package> tags. There is no
direct way to find the correct .pc file for a given assembly which
makes automatically adding a package when we do the conversion from
.csproj to a NAnt build script hard.

This patch allows to manually specify the packages needed for building
a project by adding a buildinclude.xml file to the project directory and
specifying the <pkg-references> in there. The <pkg-references> tag is
directly copied over to the generated NAnt build script.

Example buildinclude.xml file:
<include xmlns="http://fieldworks.sil.org/nant/fwnant.xsd">
	<pkg-references>
		<package name="gtk-sharp-2.0" if="${platform::is-unix()}"/>
	</pkg-references>
</include>

Change-Id: Ie47831fb89a327c58e67f0aefbdd45ac65168832
Reviewed-on: http://hydra.wycliffe.ca:9160/529
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Stephen McConnel <stephen_mcconnel@sil.org>