Fix DateChooser dialog (FWNX-817)
The calendar control on Linux has a different size than on Windows.
This means that we have to center the calendar control on Linux and
adjust the size and location of the group box and label laying behind
the calendar control (The group box and label are shown when the user
selects "BC").

To make this easier the size of the label got adjusted on Windows to
cover most of the group box and display its text centered.

Change-Id: I2f79058341f36375d3c45679e0e48e25a8182cbd
Reviewed-on: http://hydra.wycliffe.ca:9160/539
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Stephen McConnel <stephen_mcconnel@sil.org>