Don't set item height of combobox (FWNX-471)
Explicitly setting the height of items in a combobox causes problems in
a cross-platform context when the system uses different fonts.

I left two places that set the ItemHeight because they set it to a bigger
value then the default value and so it shouldn't cause any problems.

The fix for FWNX-471 also requires a mono fix.

Change-Id: Ib61091ca8bae543449722fde61790d75423d34a7
Reviewed-on: http://hydra.wycliffe.ca:9160/523
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: MarkS <marksvc@gmail.com>