Remove obsolete paracounter from DraftView
A while back the paracounters were moved from the actual view to the
view group, but the variable was left in DraftView (although never
set to anything). This doesn't cause a problem in production code
because it never gets used. It didn't get noticed in the unit tests
because they had a try/catch block around. However this caused exceptions
to be thrown which were noticeable when running in the debugger.

This change removes the obsolete m_paraCounter variable and the
(crashing) override of EstimateHeight from DraftView.

Change-Id: I9bc26a4ff4961cf4838031f26b81e54512954dba
Reviewed-on: http://hydra.wycliffe.ca:9160/561
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: MarkS <marksvc@gmail.com>