Remove unnecessary Mono.Cecil.dll
Mono.Cecil.dll is part of Mono. We should use the one included in Mono
otherwise we might run into bizarre errors. We didn't copy
Mono.Cecil.dll when building with NAnt, but when building with
MonoDevelop it got copied which lead to bizarre errors (because
obviously that was an older version).

Change-Id: Ife2818c3f4136a00c74f53edaeed99fcaf7e06af
Reviewed-on: http://hydra.wycliffe.ca:9160/568
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Brent McCarthy <brentmcc16@gmail.com>