LT-12658 XHTML contains large gaps between items
The paragraph haning indent style information was being used for styling classes other than entry. I prevented that. There still may be a similar issue with the "indent" information. If so, the appropriate place in XhtmlHelper.WriteParaStyleInfoToCss() can check the new hangingIndent parameter to inhibit that as well.

Note, putting a bogus style in a LexEntryLayout.xml layout element suppresses inheriting the indentations from the Dictionary-Normal style. I believe it uses the Default-Paragraph-Characters style.