Revert "Release rendering engines"
This reverts p4 changelist #37730 and fixes FWNX-841.

With the change #37730 applied we get a crash trying to modify a
writing system. Mono fails with a InvalidComObjectException trying
to get the IUnknown pointer for a UniscribeEngineClass object. This
happens when the temporary ws factory gets replaced with the real
one in InnerFwTextBox. Explicitly changing wsf on the UniscribeEngine
object doesn't help. Commenting the Dispose method in
PalasoWritingSystem fixes the crash.

It turns out that we don't need the fix introduced in #37730.
Running the tests with MTA as introduced in #37731 seems to be
sufficient.

This looks like there is a Mono bug related to reference counting
of RCWs but I haven't been able to put my finger on it.

Change-Id: I8cd189a5283679b58a87f77ce8b14a8e62e04701
Reviewed-on: http://hydra.wycliffe.ca:9160/604
Reviewed-by: Brent McCarthy <brentmcc16@gmail.com>
Tested-by: Jenkins <jenkins@sil.org>