Fix calculation of menu overflow (FWNX-829)
The height calculation is now done based on the monitor the
menu is displayed on and so will work correctly if the second
monitor is positioned to the right of the primary one but is
smaller.

Also calculate the overflow recursive so that it displays yet
another overflow menu if the overflow menu is too big.

A related Mono patch fixes the positioning of the menu if it
is displayed on the smaller monitor.

Change-Id: I8417878094db8bc3c79b48907be5c65fe37f0afc
Reviewed-on: http://hydra.wycliffe.ca:9160/606
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Brent McCarthy <brentmcc16@gmail.com>