Set FW_CommonAppData environment variable (FWNX-814)
On a developer machine the directory /var/lib/fieldworks doesn't
necessarily exist or is writeable which prevents FieldWorks from
running. This patch sets the environment variable FW_CommonAppData
to a subdirectory of Output_$ARCH.

It is possible to use the original /var/lib/fieldworks directory
by setting the environment variable FW_CommonAppData to the empty
string.

Change-Id: I7582d4f2d2524c8db7551ee3c804fc8fb5d206ff
Reviewed-on: http://hydra.wycliffe.ca:9160/613
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Stephen McConnel <stephen_mcconnel@sil.org>