Force nunit.exe (and nunit-x86.exe) to run under .NET 4.0 framework
This makes it easier to debug unit tests because you don't have to attach to nunit-agent.
If you need to run NUnit with 2.0 framework you'll have to use the NUnit version included
in that branch, or copy nunit.exe* to something like nunit2.exe* and in nunit2.exe.config
undo this change (i.e. uncomment the 2.0 framework).