LT-12767 Improve the receive from from colleague (obtain)option for Send/Receive and LT-12877: Reusing menu command should restore window not cause problem 1) Make cosmetic changes to fwr/Src/FwCoreDlgs/ChooseLangProjectDialog. 2) After the user acknowledges receiving the project, the Send/Receive dialog (three buttons) the FLExBridge dialog and the Open Project dialog closes. 3) If a project was selected, it opens.
I set the cursor to wait, but it seems there is too much going on for it to appear until too late.

I could not submit LT-12767 until its FLExBridge part is in the main branch. So all my issues will be submitted with this one unless the source does'nt overlap.