Bugfix in nunit.core
Fix for bug 993247 (https://bugs.launchpad.net/nunitv2/+bug/993247).
This is necessary for building on Linux. Otherwise the tests may
fail because of too many open files. This happened with FDOTests.

The dll got recompiled from https://github.com/ermshiperete/nunitv2,
tag 2.6-sil. Recompilation was done on the command line with:
build.bat release (which compiles for .NET 3.5). It was necessary
not to override the existing .config files in Bin/NUnit and
Bin/nant/bin/NUnit so that NUnit still runs under .NET 4.0.

Change-Id: Ie9bbfae9c72fa870cfe1e356332b1f131c1b609b