Replace NUnitForms with a stripped down version
NUnitForms depends on NUnit. Each time we upgrade NUnit we have
to recompile NUnitForms. However, we only use a tiny fraction
of NUnitForms. This commit adds a new FormsTester.dll that
contains the bare minimum necessary for the MessageBoxExLibTests
to work, but doesn't require the NUnit dependency.

This also requires some refactoring of the MessageBoxExLibTests.

Change-Id: I121c8becd141e8655347d3c76f8fdf3d66c85132