Provide new definitions.zip file for Gendarme
Gendarme gets the information about MonoTODO comments from
a definitions-*.zip file stored in LocalApplicationData/Gendarme.
There doesn't seem to be a way to provide a new one when Mono
got changed (other than downloading it from MoMA web page; but
the version provided there is quite old). We work around this
limitation by copying a modified zip file to that directory.

The operators Type.op_Inequality and Type.op_Equality are new
in .NET 4.0 and it doesn't seem to be clear if they got
implemented correctly in Mono. However, the MonoTODO comment
they have cause them to be reported by Gendarme which is quite
annoying.

This patch removes the MonoTODO comments for Type.op_{Ine,E}quality
from the zip file and copies the new zip file to the Gendarme
directory at the beginning of each build.

Additionally, Gendarme has a bug so that it treats
definitions-2.10.zip as being an older version than
definitions-2.8.zip, therefore the new zip file is named
definitions-2.9.zip instead of 2.10 or 2.8.1.

Change-Id: I970a56d7be1bbc710b74a11a5d08c3261521598f