Fix failing FwCoreDlgs test
The m_variantName_TextChanged() handler needs to get called
also when the selected index changes (as stated in the comment),
not only when the text changes. For whatever reason that got
removed at one point.

Another reason for failing FwCoreDlgs tests is a bug in Mono
(Xamarin-4921). The patch for that got already submitted to
mono-calgary.

Change-Id: Ie6df4097b9301d20c41bc7401712f9c06ae30f38