LT-12877 Reuse FLExBridge conflict report dialog. This fixes a pipe bug that prevented S/R Conflict report links from jumping into FW entries. Prior to fixing that, calling the user32 functions for restoring the conflict report window worked. Now no user32 function will work anywhere in the FLExBridge launcher class.
One solution according to MS is - "If you are using the Graphics object after the GetHdc method, call the ReleaseHdc method. This will release the context handle obtained by a previous call to the GetHdc method of the Graphics object." Maybe someone knows where that could be done in our code if it even makes sense.

Until the user32 problem is solved or a non-user32 solution is found, this issue won't be solved.