Fix ibus keyboard switching (FWNX-770)
This patch fixes the problem that the first keypress in a field
with ibus keyboard gets eaten (FWNX-770). It also fixes the problems
when using KFML ibus keyboards (Bugzilla #443). The reason was the
same: while we're in a field we shouldn't turn off the keyboard - this
causes us to loose the context.

Another problem was that we didn't set the keyboard for the selection
until the user clicked a second time in the field (or pressed a key?).

Change-Id: I1585c07cb1d63704c2072e21ede9a594c7e97d39
Reviewed-on: http://hydra.wycliffe.ca:9160/722
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Stephen McConnel <stephen_mcconnel@sil.org>