Improve handling of X11 connection (part of FWNX-884)
This patch improves the error handling when XOpenDisplay fails.
It also ensures that we close the display we previously opened.
Note that we don't check for error in XklEngine. This will cause
a crash when we ran out of client connections. Checking the
XOpenDisplay result doesn't help here because it only defers the
crash.

This patch also implements caching of the converted keycode values.

Change-Id: I3aec63f34926b7dd82bbd767d60575fcb3498a89
Reviewed-on: http://hydra.wycliffe.ca:9160/727
Tested-by: Jenkins <jenkins@sil.org>
Reviewed-by: Stephen McConnel <stephen_mcconnel@sil.org>