LT-13302 Browse pane goes blank when filter is released
The lex entries clerk was active and visible in the gui when the export dialog used it to publish the standard format file. Upon cleaning up, the dialog told the clerk to become inactive, which meant it could no longer be seen in the gui. So when it got to RecordList.ReloadList(), it skipped reloading.

Changes were made in XWorksViewBase.
Added private bool m_haveActiveClerk = false;
which acts as a gaurd in Dispose()
and is set in the ExistingClerk property.
The Clerk property was refactored to use ExistingClerk which is safer and allows m_haveActiveClerk to be set.