Build FieldWorks with ICU 5.0 and changes to dvcs layout
I also changed the directory name from Lib/src/icu to Lib/src/Icu4C so
that unzipping the contents doesn't end up in mixing source controlled
and un-source controlled files in the same directory.  Thus, icu can be
added to .gitignore without taking a chance that a change/addition is not
seen.

Change-Id: Ieca075f6f1575411460f4d08438fd05ca14195c3
Signed-off-by: Stephen McConnel <stephen_mcconnel@sil.org>