Use MONO_PREFIX if set
On the build machine we have mono installed in the home directory of
the build user and therefore MONO_PREFIX set. The build environment
needs to use MONO_PREFIX if it is set so that builds with mono
installed in non-standard locations will work.

Also, there is no need to call mono-real. It's better to call mono
so that it works without the wrapper script which we don't need on the
build machine.

Change-Id: Id9f30db5ed989024943dffa9de43b01b50381a34