LT-11268, Make Classified Dictionary view configurable
Also added autoexp attribute to part ref to enable item auto-expanding
in configure dialog.
Also made the name at the top of the pane match the tool name.
Made senses in ClassifiedDictionary use the Reversal mode for controlling
homograph number and sense number, and changed the dialog to reflect this.
(There is only one small label change in the dialog...all the rest is what
Visual Studio's Designer did, probably because it was opened for the first
time in VS 2010.)

Change-Id: Ie4a73a7d89cd6f9c7fbfa5bbcae4fb97db37da1e