Fix FwCoreDlgsTests
A recent change allows to build as non-admin user by putting registry
values under HKCU instead of HKLM. However, the Ethnologue class
wasn't updated an so only looked under HKLM, causing several tests
in FwCoreDlgsTests to fail on non-admin Windows and on Linux.

Also the code of the Ethnologue class claimed that the behavior could
be overridden by setting the environment variable FWROOT. This was true
only if the registry entries were not found.

This change fixes the Ethnologue class so that the FwCoreDlgs tests
succeed.

Change-Id: I182bd3a11ca3fc93ee467c121f6834806143bfe0