Get Language tests passing, upgrade to final ICU50
In order to make our ICU overrides work, I had to check in a new set
of the ICU DLLs, to handle an interface change that was made on Linux.

As long as I was doing this, I decided to upgrade to the official release
of ICU50. No changes to our patches were required, but the ICU code
had changed enough to make the patch non-automatic.

New versions of all the ICU-derived binaries, DLLs, and libraries result
from this change; also, I updated the contents of Include/unicode to match
the current ICU source. This required disabling a couple more warnings
in Include/IcuCommon.h.

I still had to disable one part of the old unit test, which tests something
we don't use in FW and so have not implemented in Icu50. I also added
a new unit test to verify that we get custom properties when iniitializing
the ICU overrides in C#.

I added two missing files to Icu50.zip.

And of course there are updated patch and zipped sources...

The current version of the patch combines a previously separate patch
to make the build unzip Icu50 to the right place, and improve error
reporting when the SIL Icu customization file is not found.

Change-Id: I24941254889e5b478f3be910d7109fea4586fb11