Fix an assert trying to focus a child pane of a MultiPane
The XML says that in the main interlinear tool, we should try to
focus the child pane called "InterlinMaster". This failed because
(a) InterlinMaster does not implement the required interface; and
(b) because the code that searches for the default pane had been
oversimplified to only check the two main child panes, not their
children.

Change-Id: I866e1060d611d7c1be9d516eaa80f667bfa4f18d
Signed-off-by: John Thomson <john_thomson@sil.org>