Don't ignore return value from nunit task
NUnit returns 0 even if there are failed tests. A non-zero exit code
means a configuration error or a NUnit crash. We should not ignore
those. It means a serious problem that justifies to stop the build.
Besides if we ignore it Jenkins won't report any problems because all
of the tests run might have passed.

The NUnit task now also collects the names of the test fixtures
that have failing tests. Those names are output at the end of a
remakefw.

Change-Id: Iac0742b2c95bb42d65ba8a51b2ff974862cb5669