Fix non-admin registration of COM objects (LT-13800)
The registration of the proxy fails if the user isn't a member
of the administrators group. However, that information is only
needed for out-of-process COM access which we don't use.

This change implements a workaround in that it doesn't try to
register the proxy dll when doing a per-user registration.

Change-Id: I525d702fa13f058f0e755fc56bf43ed0d1c0f862