Patch over FWNX-785
"Sometimes Enter appears to do nothing in the Analyze tab of Texts"
because in Mono closing a dialog activates the main window, which then
selects a control that grabs the keyboard input.  The patch explicitly
selects the sandbox after the main window has been activated.

Two other crashing bugs encountered in passing are also fixed: one due to
a null object and one due to a disposed object.

Change-Id: Ie57acb9ad3a623f921844111219e26ee960ef894