Don't unregister when building on server (LT-13800)
When building on Jenkins we're potentially running multiple builds
under the same user account. Unregistering DLLs during a build
may result in test failures from another build that tries to access
the information in the registry.

Also modified the RegFree and RegisterForTests tasks so that they
don't try to register the typelib of the Graphite.dll (which doesn't
have a typelib).

Change-Id: I3ce012447ae8db26af112673c17b8d909b75b9ee