Fix Jenkins build target
On a developer's machine we want to fail the build if a project fails
to compile or if there were any unit test errors.

When doing a Jenkins build we want to fail the build only if a project
fails to compile. Jenkins analyzes the test output and will set the
build result to 'unstable' if there were any test errors. However,
previously we accidentally didn't fail on build errors.

Change-Id: Ia98aeaa4e6c5df5dab77042510c75ddc12656454