Fix build (LT-13800)
When building as non-admin we have to register the dlls under HKLM
so that the regfree task can find them and create the proper
manifest file. This is OK to do because we redirect HKCR to a
subkey of HKCU. However, we also need to redirect HKLM because
trying to register typelibs creates some entries under HKLM. In
previous attempts we failed to do this step.

This should fix the build problems we currently have on the build
machine where the tests can't create any COM object because the
FieldWorks.Tests.manifest file is basically empty.

Change-Id: Ib1b37323f3dba41c2de07d9198f0fae74374b26d