LT-14018 fix problem resetting all homograph numbers
The code that initializes the homograph info cache was setting HN
to zero for all existing entries. Tests missed this because
typically in testing all entries are new, so this cache is
initialized at a time when there are no homographs. The problem
only occurred in real data files, where entries are loaded from
disk before the cache is initialized.

Change-Id: Ib3412d12f812f6d24a3a196280f41cfdeb64d903