Rename SharpZipLib to work for xbuild /t:localize
Building localizations on Linux/Mono fails because the old versions of
ICSharpLib.SharpZipLib stored in the gac do not work for the Zip task
used in building localizations.  (The search algorithm for Mono tries
the directory of the program, then the gac, and then the directory of
the calling assembly.)  Renaming our newer version of the DLL in the
Build directory to include the version number fixes this problem.

Cherry-picked from 5cd08e7 on develop

Conflicts:
	Build/FwBuildTasks.dll

Resolved by:
	Rebuilding Build/Src/FwBuildTasks/FwBuildTasks.csproj

Change-Id: I6e2596b4da931d60d8d743cd863078f39bfdd398