Fix build
Running tests on Linux produced missing dispose messages which causes
the build to be marked as unstable.

The disposing of command objects is necessary to avoid the mediator
being used after it is disposed in cases where a MasterRefresh or
FLExBridge related shutdown and reopen of FLEx are performed.

Fix the dispose failures in the tests by actually disposing of the
Commands in the test code.

This change is a cherry-pick/squash of the following changes on develop:
	33da59 Fix build
	393065 Revert Command dispose changes

Change-Id: I4ae84e84327331cb4f0b781a11d9b469b1c2cd43