Make ManagedLgIcuCollator reference conditional
The ManagedLgIcuCollator reference is not needed in Windows for
the Filters project (see FWNX-296). Trying to include it on Windows
results in warnings in the installer build because that assembly
isn’t there in a Windows build and the installer builder doesn’t
like the dangling reference. This is a follow-up of change
I106bfddf8c7cf8ae24a6d030b9941872227f00d6.

Change-Id: Id2bf138aa131f051c998b2cab83565da84e1fe8d