Fix build (LT-13800)
When building as non-admin we have to register the dlls under HKLM
so that the regfree task can find them and create the proper
manifest file. This is OK to do because we redirect HKCR to a
subkey of HKCU. However, we also need to redirect HKLM because
trying to register typelibs creates some entries under HKLM. In
previous attempts we failed to do this step.

This should fix the build problems we currently have on the build
machine where the tests can't create any COM object because the
FieldWorks.Tests.manifest file is basically empty.

This is a cherry-pick of 85a67f47d99cef7828436a4c49305edb00cdc590
and is needed so that building the release branch doesn't break
all builds :-)

Change-Id: I99990ad6665e7e8a03ed217bdb0ede2af6cd84e4