Add script to install dependencies on build agents
Add script and config file that allows to automatically install
the build dependencies on Linux build agents. The advantage of doing
it this way is that if we add a new dependency, the build agent will
pick it up automatically, and that we can have different dependencies
on different branches.

Change-Id: I7b2cd7d2baeb292ceea73675eda4a82119e5a482