Modify DB40 tests so the sames tests can run in parallel
This problem was described in
LT-13800 Allow concurrent builds

The tests were set up so that the projectName was always
"TestLangProjCS"
Therefore if running these tests in parallel by different builds,
the project folder on the build machine could be created and deleted
by multiple builds at the same time and clobber each other.

Therefore the changes made are designed to ensure all builds running
in parallel use unique folder names on the build machine.

Change-Id: I4df9a2726ab87e1133b635d62a76d93e50751eb8