Add build target for IDH related files
There is no MIDL compiler on Linux, therefore the COM related
header files need to be generated on Windows and submitted to
the repo. There used to be a NAnt target to do that, but that
never got ported to ms/xbuild.

This change adds a ms/xbuild target 'generateLinuxIdlFiles' that
generates the necessary files for use on Linux, copies them to the
correct place and extracts the interface GUIDs. The target
'ensureTLBsUpToDate' is now run as part of a remakefw; if a *.idh
file changed the 'generateLinuxIdlFiles' target gets run when
running on Windows. On Linux we fail with an error unless the
property buildtlb is true, in which case MIDL is run remotely on a
Windows machine.

This change also adds a new ExtractIIDs task to FwBuildTasks.dll
that replaces the former extract_iids.sed/cmd script. The reason
is that we got rid of sed.exe on Windows some time ago so we can't
use a sed script anymore (and one reason for removing sed.exe was
that having multiple cygwin.dlls in the path often causes problems).

Change-Id: I2e686e6fcbeba2bf8db50bc98c223f889789d572