LT-14154 add entry fields to categorized dic configure
Note that many new layouts are added, all with Cd (Classified Dictionary)
somewhere in their names. These are, as too often in XmlViews config,
more-or-less duplicates of existing similar layouts. The are added
so that configuring the classified dictionary will not change the
configuration of the main dictionary.

The change to pass the Caller in XmlVc is needed to support the more
complex generated parts with the added layer of indirection to the
OwningEntry. For example, it gives the configureMlString code access
to the surrounding PartRef with the writing system info.

Also fixed a problem Gendarme apparently missed earlier but is now
complaining about...

Change-Id: I7d7158cd02602057041970f61ea0632282247556