LT-13906 clean up safe-XML usage in Lift merger
I am assuming (and in many cases have confirmed) that the text that
comes in, mostly as LiftMultiText, from the Palaso lift parser, is NOT
'safe' XML (i.e. with &, <, etc, excaped). In most cases we do so as
soon as we get it.

I assume anything being put in FLEx data structures like TsStrings, and
certainly in FDO objects, needs to be decoded. Simple comparisons with
internal strings also require normalization.

Change-Id: Iffaeaaf324c4ea58d2bbec336087c29c6d25b182