Fix comparison to allow stable RBTree insertion/retrieval
In the Mediator, colleages are saved in a dictionary, for which Mono
uses a RBTree internally.  The comparison function defined uses the
GetHashCode() method on the objects implementing each colleague.  It
could happen that a colleague could get stored in the dictionary that
could never be retrieved after other colleagues were stored in certain
orders due to arithmetic overflow in the comparison function.  This is
the root cause of the bug reported in FWNX-896.

Change-Id: Ie7aee60b089db55acb14735605bac0cc548a5beb