Fix localization (LT-14280)
A recent change removed obsolete files from Bld. It wasn't obvious
that one of the files was still needed for localizations.

This change moves the LocalizeResx.xsl to Build so that it is
together with the files that need it.

The change also includes an updated LocaleStrings.exe that retrieves
the FW version information from GlobalInclude.properties.

And finally it fixes the LocalizeFieldWorksTests.

Change-Id: Iffa3c3c1d20bef34d9dd3adcbbad31a768531d9c