Use FW_ROOTCODE to find strings-en.txt
This fixes FWNX-938, FWNX-939 and possibly other issues.  When the C++
string resource file cannot be loaded, this can cause BeginUndoTask to
be called with empty strings, which throws an exception and leaves the
program in an unstable state.  (This bug was very hard to track down
because developer builds always found the string resource file in
Output/Debug, the universal working directory used by programmers when
running the program!)

Change-Id: I7ccead524c5a942a0ad3cd32c2bfd504ee182eee