Circumfixes in Hermit Crab (LT-14321)
Any inflectional affix marked as a circumfix was not being treated as
filling an inflectional template slot.

Also, if an affix template had a slot for the prefix portion of a
circumfix but did not have the suffix slot for the suffix portion, could
get a yellow crash (index out of bounds where it expected to find the hvo
of the missing suffix slot).

Change-Id: I14afab906edf06fcafdb154cc81e8a9b0f0188df