Fix failing tests when default backup directory was set
FwUtilsTests failed when a default backup directory path was set.
We now use a special registry key for the tests so that we don't
care about the live settings. This also allows to enable the
backup directory test on Windows again which got commented out for
similar reasons in LT-13019.

Change-Id: I33fa19a4b088d0b3b439cf12803eaa5c64c55cbc