Prevent growing fwlayout files
This fixes FWNX-1057 "layout configuration files can grow with duplicate
entries".  I suppose you could argue that this reflects a Mono library
bug, in that removing an child element from an XmlElement invalidates
(and thus terminates) an iterator for an XmlNodeList involving child
nodes of that element.  But it's an easy fix in FieldWorks code.

Change-Id: I83a09f7c8ec7c6cbd6b39cf51017090ece9bd9c7