Improve libhunspell handling on Linux
Mapping libhunspell to the actual library in the config file has
the advantage that it can be changed without recompiling the code.
The <dllmap> entry in the config file will be ignored on .NET (see
http://www.mono-project.com/Interop_with_Native_Libraries).

The original motivation for this change was that we get test
failures on Lucid where we have libhunspell-1.2. Unfortunately it
is not possible to override dllmap entries. On an end user's machine
we'd have to modify the config file (probably when building the
package). On the Lucid build agents we use a hack and create a
symlink to libhunspell-1.3 :-)

Change-Id: I3278bb364bc2bfaea5dcbb06cf0ae6053046270a