Fix LT-14350 and one other issue
Flex has two ways to connect to Flex Bridge, one of which does not
lock down Flex, while Flex Bridge is running. That method had been
used for the two 'view message' options, but now it is also used
for the "About Flex Bridge" and "Check For Updates" menus.

The fix for LT-14350 may not work for cases, such as obtaining a project
via the Welcome dlg. A Mediator instance is required to save which
bridge was last used, and there may not be a Mediator present in
that context, or in some another context really early in the
start up system.

Change-Id: I8930de8db47b1cbb092021fa14a02ba98544cbdc