LT-14436, Localize build task needs to quote product version
Most of the change is just so we can test it. The only significant
difference in the actual code is putting Quote() in the line
builder.Append(Quote(productVersion)).

I haven't been able to do a real test of this because the current
develop build doesn't have a problem product version.

Change-Id: I8a734494b766cb1d6c06e584d5c34b7d75a26a1e