Fix reporting of failed tests
When we get test failures we don't want to fail the build, but
add the failed tests to a property that we report at the end of
the build. This allows Jenkins to report the build as unstable
instead of failed.
(cherry picked from commit dbd13cc0a9c9a65d9fc211f75f86c94118002bdc on
develop)

Conflicts:
	Build/FwBuildTasks.dll
	Build/Src/FwBuildTasks/TestTask.cs

Change-Id: I49f9832b484cc063cefb8b9db1f127e607a26dfa