Fix problem with non-existing linked files folder (FWNX-1005)
Another attempt to fix the problem when the linked files folder
doesn't exist. This can happen when the project gets moved between
machines. The previous attempt caused a dialog box to pop up
recursively causing it not to show anything. We now check for the
existence of the linked files folder when we open the cache.

Change-Id: I20756db7a5cf570d2c47e118c4b52979bba3cf66