Fix loading bug (hangs forever) found on Linux
When investigating FWNX-1078, the test project uploaded for the issue
hung when trying to restore/load it.  Both the "Domain Loading" thread
and the main thread were trying to store information in the same Dictionary
object at the same time.  Although the problem has not been reported on
Windows, the fix seems reasonable to apply across the board.

Change-Id: I7a7b65c282bb87fc3ee16423ce0e0e3154a33ab7