Fix failing tests
A recent update of Mono caused two failing tests on Linux
caused by lazy creating the document stream. Setting the
XmlResolver to null caused a crash because it then couldn't
create the stream anymore. Workaround is to pass the stream
to the Load method of the XmlDocument.

Also added the two .xhtml files used in the tests to the test
project.

Change-Id: Iaca87512750087b19da8f72fee379fc0672b4645