Changes to allow Linux use gecko for OCM helps
Implemented the GeckoWebBrowser in the control to allow Linux access to
the extracted htm files from a combined OCM/OCMFrame.chm. Linux is not
able to reference the htm files directly from the chm format. The htm
filenames and hrefs needed to be converted to lowercase because of case
reference problems.

This fixes the display for OCM help pages in the Choosing
Anthropology Categories Dialog. (FWNX-464)

Change-Id: Ia59ac946101a18efe9ba867c073c552d9860277f