LT-13883 Cleanup after removing all inflection features
Note: I haven't figured out a way to do unit tests for these changes.
One is merely the removal of an unhelpful special case in a method
that also launches a dialog, making it hard to test.
The other is a minor change in a class that is a real control, and would
require significant re-architecture or test-only methods to be testable.
Doing that feels more risky than this very minor change.

Change-Id: Ifd4d1e4c79cb37b8a1b15469bd75230d7cb67a39