Pop up message box if registry entries don't exist (LT-14462)
The previous version of the fix for LT-14462 prevented FieldWorks
from run if the registry entries existed under HKCU but not under
HKLM. However, with the current FW version it is possible to run
with only HKCU registry entries. This patch improves the check for
the missing registry entries and will only pop up a message box if
both HKLM and HKCU registry entries are missing.

Cherry-pick from 9d595f on release/8.0 branch.

Change-Id: I3d92a335c4a16a2104ba827a035d3bef753066b1