LT-14574, choosing current sense should update MSA
In the interlinear focus box, when the user pulls down the chooser
on the Lex Entries line and chooses the selected (current) sense,
usually nothing needs to happen. But if the sense has an MSA and the
morph bundle does not, we should treat it as choosing a new item,
so the analysis will get adjusted.

Change-Id: Id52bc05d33b0501d5c94710be21a2b4630133736