LT-14558 use template folder for missing writing systems
This has two parts:
- whenever we create a temporary PalasoWritingSystemManager, we need
to tell it where to go for templates for writing systems not in the
global store
- when creating a new project, make sure all the WSs from the template
folder are offered as options.

Several tests had to be changed because they now find the Template
folder versions of writing systems. This means that in English,
single-quote and hyphen are word-forming. It also gives a different
name to Chinese.

Change-Id: I0b08b376a1db215ee3a48827feb11d4da09a5401
Signed-off-by: John Thomson <john_thomson@sil.org>