LT-14562: change message when unable to restore old project
Pre-7.0 projects can't be restored (if they don't have XML already)
on non-Windows operating systems or versions after Windows 7
(because SqlServer 2005 is not supported). This change puts a different
suggestion in the dialog for such users, and hides some irrelevant stuff.

The only real change in Resources.resx is adding the new
kstidCantMigrateWrongOS string. All that changed in the designer is
giving member variables to some additional controls we need to be able
to hide or otherwise manipulate.

Change-Id: I388899e6a4bae21f08ea66f4ae32dd1abb5d864f