LT-14610, imported FlexText should appear after import
Proved to be a very subtle bug introduced when we made Texts
unowned. We attempted to create a PropChanged notification for
new texts, and succeeded for ones created normally. However, when
importing FlexText, we want to restore the original guid (I think this
is for ELAN interchange). The Create method used to do this was not
calling the method than enables the PropChanged.

As far as I can tell, Text is the only class in the system that can
be created unowned with a fixed guid (the one combination of
circumstances where this hook was not called).

Change-Id: Id19a574ee7d6c88e98c61c4f7f55d62d11f56ab7