LT-14066, Hide "will continue" for non-fatal error reports
This dialog is used for various non-fatal error reports, such as
Help/Report a problem, where the main button text is changed to "OK"
rather than something like "Exit now". In the fatal case, holding
shift down is a way to attempt to continue, and a message indicates
this. But it is distracting while typing into the dialog, and
inapplicable to non-fatal errors. So we have been asked to change it
to only appear when "OK" is normally going to kill the program.

Change-Id: Id3f14c4e5e9bf4f376e292c134f93d976ccc239c