Make ProjectLocationSharingDlg's purpose clearer
LT-14483
Project Locations and Sharing path does not match reality or the
FieldWorks Project Properties General tab

When examining the code it was discovered that the dialog is
displaying the correct information. To help avoid confusion
it was decided that showing the current project's Path would
help.  Therefore the changes to the code reflect that change.

Change-Id: Ic78a5739cab091982eaecf8747804e9f29fa3f38