FWNX-787 Make ConcDialog record status visible in Mono
In the ConcordanceDialog there is a StatusStrip with three items.
The last is m_toolStripRecordStatusLabel, which displays information
about the selected item in the dialog (1/3, or No Records).
Previously, the widths of the three controls were 200, 96, and 96.
With a default 2 pixel GripHandle, the right of the last item comes to
399. Since the StatusStrip is 400 pixels wide, it just barely fits...
on Windows. In Mono, for a reason I have not determined, the
DisplayRectangle of the StatusStrip (in which the controls are to be
laid out) is smaller...385 pixels wide on my system. (One theory:
leaving space for a window resizing grip?) The StatusStrip
decides that the record status item does not fit, and hides it.
The only solution I have found is to make it smaller. It seems to be
still plenty big enough for "No Record", probably the longest string
it ever shows.

Change-Id: Ib0195f21a4f7a001c9c4a1cdcffe0ca6029d25ec