Bug LT-14560: Parser Crashes in Rare Circumstance
The crash occurs when parsing any one of three words in a particular
9505-word dictionary.  The crash was caused by an index out of bounds
exception when trying to access a lexical entry, specifically, the
morphosyntactic analysis for the preceding sibling (the second element
in a one-element array).

Update the XSLT script that generates this index to account for
0-indexing.  Update "unit" test to reflect this change.  Add some
minor comments in the process.

Change-Id: Ib91e7f8d45d901b49e29626bf5a3f89250256328