LT-14628 Cancel of importing Flextext doesn't work
Now if the user clicks the cancel button the import will abort.
It also shows the text with a writing system that
is not currently part of the language project so they can make
a more informed choice.

The DialogResults (OK/Cancel) are returned instead where
OK indicates the writing system will be added to the project
and all other texts with that ws will be accepted.
Cancel results in the import being Aborted.


Change-Id: I14e7bdd6b47f6c2f2cad9a137e5b5d43e039e59c