FWNX-1017 Move select all button into ScrollContainer
On Mono, despite being supposedly in front of the scroll container,
the select all button was sometimes hidden behind it. Putting it
inside prevents this.
In any case, this button is supposed to scroll with the other things
in the scroll container, so it belongs inside it. (This fixes
another bug, but I don't know whether it has been reported.)

Change-Id: Id3e338dd18f2cc664abb4f233aaa679e8bf43c9d